requirements.txt
autopep8==1.7.0
pycodestyle==2.9.1
python-slugify==6.1.2
pytube==12.1.0
text-unidecode==1.3
toml==0.10.2
